Defer template button creation till needed
This is needed because it delays application startup and every unit test that creates an application was waiting for this code to gather all templates which required opening every engine, project and gem .json file looking for templates.  This should speed up each application test by a couple seconds and save several seconds overall.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>