Make profile the default configuration in Android. (#10191)
Running `deploy_android.py` just generates the gradle solution, but it's not until the first time it's opened in Android Studio that it will run cmake configuration, and since the default configuration was debug it did the cmake for debug. Then when switching to profile it runs cmake configuration again, then you can build the project.

The most common configuration used is profile. So by setting it as default, the first time the Android Studio is opened it'll will configure cmake for profile and that way you only run cmake once and not twice for the most common case.

Signed-off-by: moraaar <moraaar@amazon.com>