Asset Processor Startup time Optimization
The Asset processor was previously rehashing every single file,
every single time it started, even if the file time on the file matched
exactly the file time it saw last time, and it had a last known hash
saved in the database.

This change adds a step which pre-warms the file info cache up with
the data from the database and the file scanner, which drastically
improves the startup time of AP when most assets have not changed.

(Debug improved on this mac from ~5 minutes, down to ~10 seconds).

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>