Fixes a crash but that was happening on linux only
This crash bug should have been happening on all platforms.
ProcessCommunicatorTracePrinter's constructor is passed a pointer to a
ProcessCommunicator* .  It caches this pointer and makes calls to it
from that moment on.

If ProcessCommuncator* is deleted before the
ProcessCommunicatorTracePrinter is, its possible for calls to be made
to the dead pointer.

Whenever ProcessCommunicatorTracePrinter is used, it needs to be
destroyed before the ProcessCommunicator* that was passed into its
constructor.

Signed-off-by: lawsonamzn <70027408+lawsonamzn@users.noreply.github.com>