add refresh button, init when active screen only
Avoid the performance hit of rebuilding the Gem Catalog and Gem Repo screen every time the Update Project screen is shown and don't assume the catalog changes every time we visit it - usually it doesn't.  But give the user a refresh button in case they know it has changed and want to load latest from disk.

Signed-off-by: Alex Peterson <26804013+AMZN-alexpete@users.noreply.github.com>