GCC static_assert fix
The AZ_VA_NUM_ARGS macro was triggering the static assert when counting an empty argument list in variadic.h

The fix is to use detect support for the C++20 `__VA_OPT__` macro and use it to optionally add a comma instead of using GNU CPP `##` extension
https://gcc.gnu.org/onlinedocs/cpp/Variadic-Macros.html

Since the minimum version of GCC that O3DE supports(9.4) that has support for the `__VA_OPT__` macro in C++17 mode, this fixes the issue when counting variadic arguments.
Clang is using the `__VA_OPT__` supported codepath as well in C++17.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>