Modify CryLog to use FileIO system, that way logs will work when using VFS. (#10484)
Signed-off-by: moraaar <moraaar@amazon.com>

## What does this PR do?

Modify CryLog to use FileIO system, that way logs will work when using VFS.

It also fixes a bug with the game log using the launcher where if the game exited by the command `quit` it was closing the log and then some teardown code will log more, which caused to open the log again and that causes to wipe the log (because the implementation is to wipe the log on first open). Now on exiting, the log will be flushed but not closed, so all the teardown code can also print to the log. The log will be closed on destruction of the class.

## How was this PR tested?

- Open editor, enter game mode and exit. The editor log is created successfully
- Do it again and checked the backup is created.
- Open launcher and enter command quit. The game log is created successfully.
- Open launcher and enter pressing X. The game log is created successfully.
- Run Android using VFS, the game log is correctly created in PC's `PROJECT\user\log` folder.