[Terrain] Fixing property tree editor to not explode when a editMetaData has no name (#10553)
[Terrain] Fixing property tree editor to not explode when an EditMetaData has no name

Previously if an ElementData had no name, the PropertyTreeEditor could explode when building a node map. Since it's possible for components to provide their own ElementData using SetDynamicEditDataProvider, there's no guarantee that the ElementData will actually have a name. This change makes it so the node will be skipped if it has no name, and a warning will be generated that helps guide people to the problem.

This change also adds behavior context definitions for a couple terrain classes that were involved in the original bug. While they didn't end up being required to fix the bug, they're still helpful to have.
____________________

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>

* Moved the warning out of the if

Signed-off-by: Ken Pruiksma <pruiksma@amazon.com>