[Terrain] Gradient dirty region support (#10674)
* Add an API to refresh a region of terrain, instead of having to refresh an entire terrain spawner.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Removed unused alias.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Added a new optional OnCompositionRegionChanged() API to DependencyNotificationBus.
This new API lets a component announce that a specific region is dirty, as opposed to the previous OnCompositionChanged() API that simply marked the entire entity as dirty.
The new API is opt-in for listeners. By default, it will route back to OnCompositionChanged().
The Surface Mask Gradient component has been modified to use the new API.
The terrain components have also been modified to listen for the new API, and only update a part of the terrain when received.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Fix transform bug in the gradient sampler and add unit tests.
The gradient sampler was transforming in the exact inverse of what it was supposed to. This has been fixed, and unit tests have been added to verify the results.
In addition, a convenience method has been added so that the gradient sampler can transform an AABB by the gradient transform so that a gradient's dirty region can be appropriately transformed in world space.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

* Introduced custom notification functions into Dependency Manager.
This allows components to take region-based notifications from their dependents and modify them before passing them forward. For backwards compatibility, region-based notifications get turned into full-entity notifications.
Terrain Height/Surface components and Surface Mask Gradients have been modified to opt in to region-based notifications.
Mixed Gradient component has been modified to take a dependent gradient region and transform it via the gradient transform before passing it forward.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>