SerializeContextTools `dumpsc` support for loading project gems
The `dumpsc` command for the SerializeContextTools has been updated to support loading a project's Editor gem modules if the --project-path parameter is supplied.

Changed the --output argument for the `dumpsc` command to output relative to the current working directory, instead of relative to the project path.
When the project path wasn't specified this would output files within the executable directory instead of the current working directory.

Added an "IsComponent" field to the dumped SerializeContext content.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>