Removed the HphaSchemaErrorDetection.cpp compilation
The HphaSchemaErrorDetection.cpp file is being removed with the
Allocator availabilty work and because it was including HphaSchema.cpp
directly within it, it was causing namespace issues compiling non-fully
qualified AZStd code.

For example the `initializer_list` class is aliased in the AZStd namespace. So when `initializer_list` is specified instead of `AZStd::initializer_list`, it would cause a compilation issue because the `Internal::AZStd::initializer_list` symbol is not available because it wasn't aliased inside of the Internal namespace.

Pretty much the issues that occurs with including a cpp file with a namespace is as follows

1. At global namespace scope a file let's say such as <AzCore/base.h> is included.
   That aliases `std::initializer_list` into ::AZStd namespace
2. Later on in the Internal namespace HphaSchema.cpp was included, which includes <AzCore/std/containers/unordered_map.h> for the first time.
  That unordered_map.h would try to include <AzCore/base.h>, but it can't because it was already included outside of the Internal namespace
3. The unordered_map class implementation occurs in the AZStd namespace(which is now Internal::AZStd, because its header was indirectly included within the Internal namespace) references the symbol the `initializer_list` without AZStd in front of it
4. Because there are symbols with the names of `Internal::AZStd::initializer_list`, `Internal::initializer_list` or `::initializer_list`, the compilation fails.

TLDR: The inclusion of a cpp file with other includes inside of a namespace, was putting the onus on the included files to deal with an issue of fully qualifying names, when the issue was caused by the inclusion of the cpp file.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>