Removed push_back/push_front/insert extensions which took no parameters.
The call sites have been replaced with calls to emplace_back or emplace_front.

Corrected return types for AZStd insert() functions to return an `iterator` instance for every insert function which accepts a `const_iterator` position parameter as many of those functions were incorrectly returning void.

Updated AZStd::deque emplace_back/emplace_front function to correctly return a reference instead of void.
Updated AZStd::ring_buffer extension to support emplace_front and emplace_back functions.
Updated AZStd::forward_list to support rvalues and emplace_front/emplace_front functions. AZStd::forward_list previously didn't support any rvalue functions for move.

Since AZStd::forward_list is only used in unit test it is not much of a problem.
It is also has a lot of limitations as compared to AZStd::list which makes it not quite useful to ever use in the codebase.

Finally removed the swap() function extensions which accepted an rvalue reference.
They have been replaced with assignment to a default constructed instance, which what the swap function was swapping with it anyway.
The operation to swap with an rvalue is really just a move operation since it is stealing the values from a temporary.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>