Fixed native visualizers for the AZStd forward_list and fixed_list classes
This fixes the forward_list iterator types as well to be casted to the correct forward_list::iterator type.
Previously it was being cast to the list::iterator type which is not compatible.

Added a [next] entry in the forward_list iterator entry that allows traversing the iteratorxs entries in the debugger with the list node values shown
Also added a [next] and [prev] entry in list iterator netry that allows traversing the iterator entries

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>