[Terrain] Macro materials store data based on terrain rendering distance instead of world bounds. (#10785)
This PR changes the way macro material data is stored so that only the area around the camera needs to be resident in memory. The amount of memory used for macro materials will be based on the terrain rendering distance. Previously macro materials were stored in a grid across the terrain world, which required explicit x/y bounds on the terrain. This removes the need for macro materials to have any such bounds.