[Terrain] Sorted macro materials
This adds the ability to sort macro materials by giving them a priority, so their order is consistent. The priority can be positive or negative to easily allow users to make macro materials more or less than the default. This change also fixes a bug where if more than 4 macro materials existed in a tile and one was removed, the 4th spot wouldn't be filled with an available candidate macro material.

This change also makes AZ::RHI::Handle constexpr friendly

Internally, much of the TerrainMacroMaterialManager was refactored to reduce code and data duplication and change from using raw numeric indices to handles.