Updated the SettingsRegistry Notify event handler signature
Instead of separate arguments being supplied for the JSON key path and
the JSON key type, a structure of NotifyEventArgs are now supplied to
the handler.
This is to allow for extensibility of the Notify Event Handler
parameters in the future.

Furthermore the NotifyEventArgs contains an additional member for the file origin path of the
file being merged into the Settings Registry.
This requires the Settings Registry to now track a stack of files being
merged.

The Settings Registry Merge event handler signatures as been updated as
well to supply a MergeEventsArgs struct to the handlers.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>