Updated the SettingsRegistryOriginTracker AddOrigin function.
It now uses a do .. while loop + a call to Dom::PrefixTree::VisitPath to
validate if there are any parent settings key that is missing from the
origin tracker.

Updated the Settings RegistryOriginTracker RemoveOrigin function to not
remove children when it removes prefix.

Various formatting fixes to be in line with the O3DE Coding
standard(curly braces on a newline, removal of unnecessary semicolons,
etc...)

Adding a googletest printer for the SettingsRegistryOrigin struct to
pritty print the content of the struct when it fails a google test
expectation.

Remove the SettingsRegistry PreMerge/PostMergeEvent handlers from the
SettingsRegistryOriginTracker as it is unneeded now
The SettingsRegisty NotifyEvent now suppliesa file origin path

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>