[Terrain] Macro material order determinism (#10893)
[Terrain] Macro material order determinism

This PR makes macro materials fall back to the entity ID for order if their priorities are the same. This isn't meant to offer true stability across runs, but keeps the macro material from looking broken in a session. Currently it's possible for different terrain tiles to have different macro material orders for the same materials if the priorities are the same. With this change, all the tiles will have the same order.