Removed job grace period and fingerprint-changing wait.
This block of code was really not serving any purpose.  If the fingerprint of a file changed, AP would pick up the file change and trigger another job anyway.  rccontroller would attempt to cancel the already-running job but this would fail since there was no listener set up at this point.
By removing this block, AP is much more responsive to rapidly changing files with no apparent downsides.

Note that testing also showed that a file which was opened for write with no shared access (exclusive write) would fail at the CreateJobs stage since the file can't be opened to even create the job, further making this code useless since it wasn't even possible to get to this code while the file was still locked.

Signed-off-by: amzn-mike <80125227+amzn-mike@users.noreply.github.com>