[Terrain] raytracing no longer uses world bounds (#10895)
Previously the terrain ray tracing mesh used a fixed number of points spread out across the whole terrain world to make a low resolution ray representation of the terrain to function as a starting point for future work. However, we're wanting to remove the dependency on x/y world bounds since it isn't really necessary - what's important is being able to draw the area around the camera to some distance, but it's ok if terrain data exists beyond that distance. This requires that the ray tracing terrain mesh no longer uses the terrain world bounds. For now I changed it to create a mesh centered around the origin, but lined up with the query resolution. This means that the ray tracing mesh will match the highest LOD rendered mesh, but only in the area around the origin.