NameDictionary non-singleton support
Add support to the NameDictionary class to allow for local instances to
be created in addition to a global instance.

Switched the NameDictionary to register an instance pointer with an
`AZ::Interface<AZ::NameDictionary>`

The NameDictionary::IsReady and NameDictionary::Instance function now
returns the instance registered with the AZ::Interface

Changed the NameDictionary::Create function to not create a static
instance if it a NameDictionary instance is associated with an
AZ::Interface. Also simplified the logic to just invoke
`LoadDeferredNames` to link static Names into that NameDictionary instance
deferred head

Updated the Internal::NameData class to store a reference to the
NameDictionary that created it.
This allows the Internal::NameData class to avoid accessing the global
instance of the NameDictionary when releasing the Name.
This also has a second benefit in that the NameDictionary destructor can
now invoke `UnloadDeferredNames` as it doesn't need to be registered
with global instance in order for the NameData to access it.

Added a ScopedNameDataWrapper as a private member of the NameDictionary
class which updates any Internal::NameData that it creates with a
reference to itself. When the NameDictionary release it's reference to
the Internal::NameData object, it resets the Internal::NameData
reference to the NameDictionary to nullptr

Rollback the change to have the IsReady function to create the
NameDictionary
automatically(https://github.com/o3de/o3de/blame/205bb2d1fe351471c03843e0c2f4487ea279a9b0/Code/Framework/AzCore/AzCore/Name/NameDictionary.h#L60)
It is completely illogical as by default the act of checking if the
NameDictionary is ready would create it.

Moved the creation and destruction of the NameDictionary in the ComponentApplication, to the ComponentApplication constructor and Destructor so that it is available as early as possible.
The SettingsRegistryOriginTracker uses AZ::Dom::Path which requires a
NameDictionary to exist.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>