Fix for cycle dectection logic in the o3de manifest.py
The get_gem_external_subdirectories function wasn't removing the path
that completely visited from the set of visited gem paths when detecting
a cycle.

Also changed the visited gem `set` to a `list` to maintain insertion
order so that the error output show the visit order that led to the
cycle.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>