Fixed the AZ::IO::Path RootPath function
The RootPath function was not returning the root directory part of the
path if that part didn't match the preferred path separator.

Also fixed the PathDecomposition functions to correctly propagate the
preferred separator of the path being decomposed to the returned
PathView.

Updated the ComparePathSegments and HashSegments functions of the
PathParser to be constexpr again by using std::is_constant_evaluated
intrinsic to branch the logic on compile time vs run time.

Added test to validate the Path Decompisitions functions when using a
path instance with the windows path separator set as the preferred
separator, but the RootPath itself containied a Posix path separator

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>