Updating the AZCore Utils functions
The Utils function which retrieves paths from the SettingsRegistry have
been updated to accept an optional pointer to the SettingsRegistry to
use to lookup the paths.

Added a GetGemPath function which can query the SettingsRegistry for the
absolute path to the gem root by supplying the gem name.

Updated the ReadFile utility function to support reading into an
`AZStd::vector<AZStd::byte>`

Updated the WriteFile utility function to write from an
`AZStd::span<AZStd::byte>`

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>