Implement container Card
It's a new style for card that is meant for "sections" in property
editors for grouping things.

To implement the underline in CardHeader a new internal widget was
introduced, a simple single-colored Rectangle. It wasn't feasible
to implement the underline with stylesheets because:
 * the color may change at runtime so it can't be explicitly set
   in the stylesheet
 * ::setStylesheet() would conflict with what's happening in
   Card::polish()

Container card header styling got partially decoupled from the Card
itself because CardHeader will be later used stand-alone (see #10960).

Fixes: #10958
Signed-off-by: Miłosz Kosobucki <milosz@kosobucki.pl>