Made PreviewRenderer disable the ImGuiPass to avoid irrelevant input handling.
Debugging some ImGui issues was difficult because multiple render pipelines were handling mouse input, but only one of them actually used ImGui features.
Note that I had to change Pass::OnHierarchyChanged and Pass::OnOrphan to be protected instead of private. As private functions, they could be overridden in base classes but could not call the base implementations.
I also changed Pass::IsEnabled to return false if any of its parents are disabled, to reflect the fact that it won't be executed in that case anyway. Normally I wouldn't like the idea of SetEnabled and IsEnabled having different semantics, but that's already the case, because there are pass subclasses that override IsEnabled with custom meaning.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>