Fix VK_EXTENSION_SUPPORTED macro by checking the runtime variable of glad context (#11151)
Signed-off-by: moraaar <moraaar@amazon.com>

## What does this PR do?

Fixes `VK_EXTENSION_SUPPORTED` macro to check at runtime if a vulkan extension is supported or not.

Glad vulkan checks each extension availability at runtime (when loaded) and saves it in global int variables `GLAD_VK_EXT_extensionname`. PR https://github.com/o3de/o3de/pull/10118 changed this so it checks if a macro `VK_EXT_...` is 1, which it always is, but it's not checking at runtime if the device/instance has support for the extension. So what's happening is that on android it now thinks extension `EXT_debug_utils` is available, which is not, and then it crashes when calling `CreateDebugUtilsMessengerEXT`.

With the new glad vulkan header introduced in PR https://github.com/o3de/o3de/pull/10118 instead of checking `GLAD_VK_EXT_extensionname` it should be checking `context.EXT_extensionname` int instead.

## How was this PR tested?

Run default level using vulkan rhi on pc and android.