[Terrain] Fix for macro materials when terrain render distance > 8km
This PR fixes an issue with macro materials on large terrains, when the draw distance is greater than 8km causing the uint16 tile index to overflow. This change also removes noisy warnings from feature processor factory (sometimes these functions are called with the expectation that they may fail) and an unneeded include.