[Environment] Allow for explicit instantiations of EnvironmentVariables with move-only and copy-only types
This disables the `Set()` functions of an `EnvironmentVariable` if a type
doesn't support move assignment or copy assignment. This is required for
explicit instantiations of the EnvironmentVariable template with move-only
or copy-only types. An explicit instantiation of a template class will
instantiate all methods of that class, including ones that are never used,
so even methods that are not used must compile.

Signed-off-by: Chris Burel <burelc@amazon.com>