Fixed overwriting of an install CMakeLists.txt with install permutation content.
The generate empty CMakeLists.txt would override the CMakeLists.txt
generated in the `ly_setup_subdirectory` function causing the include
command within it to be removed.

Added a global property of
`O3DE_SUBDIRECTORY_${real_external_subdirectory_path}` for each External
Subdirectory to provide a quick way to lookup if a directory is an
external subdirectory.
This is used to prevent the `ly_setup_install` function from invoking
add_subdirectory on an external subdirectory that would be added via
visiting the "external_subdirectories" paths in the SDK layout
engine.json.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>