Fixed race condition in the NameData::release function
The NameDictionary
`NameTest.ConcurrencyDataTest_EachThreadRepeatedlyCreatesAndReleasesOneName_NoCollision`
was sporadically failing due to a race condition in `NameData::release`,
where the m_useCount member could be decremented to 0, followed another
thread invoking `NameDictionary::FindName` to locate that NameData
instance and initialize an AZ::Name with it which increments
the m_useCount back up to 1 and then that same thread can
have that AZ::Name instance go out of scope causing the m_useCount to
decrement back to 0.

What then occurs, is that one of threads can invoke
`NameDictionary::TryReleaseName` which deletes the name data instance,
while another thread could attempt to access the deleted
`m_nameDictionary` member of the NameData class.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>