Replaced the LogNoFormat function with LogWithCallback
LogWithCallback is used to allow the caller to supply a callback, which accepts an `AZ::IO::GenericStream` to be writing to with the message content to log.

Renamed the LogPlus, LogToFilePlus and LogToConsolePlus function to LogAppendWithPrevLine, LogToFileAppendWithPrevLine and LogToConsoleAppendWithPrevLine to better clarify those functions intent.

Moved the Timestamp logging lambdas in `LogStringToFile` to a free function to allow for reuse in the `LogStringToFileWithCallback` lambda inside of `LogWithCallback`

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>