Added a no-index GetSwitchValue overload to AZ::CommandLine
The new GetSwitchValue overload which doesn't accept an index returns the last occurance of a command line option.

This helps to prevent user from coding to retrieve the first instance of a switch on a command.
Ex. `commandLine.GetSwitchValue("<switchName>", 0)`.
CLI argument parsing uses the last instance of the parameter as the value.
i.e `command -s 10 -s 20`, for the option for 's', if only one value is used, it should be the last one of 20

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>