Fix for polygon prism single-thread deadlock. (#11378)
The polygon prism has a complex scheme for managing its shared mutex so that it doesn't get double-locked on the same thread, which can cause a deadlock. However, the scheme was getting bypassed when the shared mutex was passed directly into the IntersectionDataCache. It needed a bit more logic to *not* pass the shared mutex in the case that it was already exclusively locked.

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>

Signed-off-by: Mike Balfour <82224783+mbalfour-amzn@users.noreply.github.com>