Settings Registry Visitor API Updates
[[Breaking Change]] The Visit functions on the Settings Registry now supply a struct with the arguments about the field being visited.

This allows future updates to Visit API without changing the API in the future.
The new struct now supplies the `SettingsType` value for the field type, instead of `Type` allowing for querying the signedness of an integer type.
The struct also supplies a reference to the Settings Registry being visited, alliviating the need to bind the SettingsRegistry to the visitor or functor, supplied to visit.

Greatly simplified the PlatformConfiguration parsing of the settings registry for AssetProcessor settings, the LocalBookmarkLoader for the editor bookrmark settings and the StreamerConfiguration for determining if a drive is in use by O3DE

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>