Prevent assert from occuring when logging messages >512 chars on a non-main thread
Previously as of commit
https://github.com/o3de/o3de/blob/ef9778ffd9a35f07569e6f52809efbfd59bfbc71/Code/Legacy/CrySystem/Log.h#L112
the `SLogMsg` would silently truncate log messages of greater than 512
characters if they occured on a thread other than main.

With the change in the latest commit
https://github.com/o3de/o3de/blob/900bd390b9b11c83a77ac3de651dc9eb3d897270/Code/Legacy/CrySystem/Log.h#L107
to switch to using a `fixed_string<512>`, an assert occurs when copy
more than 512 characters to the message buffer.

This change now uses an `AZStd::variant<AZStd::fixed_string<512>,
AZStd::string>` to populate the fixed_string up to its max size of 512
characters before moving to use a heap allocated dynamic string if the
message is more than 512.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>