Add new class for tracking allocations
When allocators are constructed before main starts, there is no easy way
to configure their Descriptors. To simplify the dependencies of the
allocator system, the Descriptors will be removed. One thing the
Descriptors provide is a way to configure if the allocator instance will
track its allocations or not. Once Descriptors are not able to provide
that information, we need a different way of specifying which allocators
track their allocations, and which ones do not.

This new class will be used to define that relationship at compile time.

Signed-off-by: Chris Burel <burelc@amazon.com>