Material Editor: Fixing access violation in EditorModeFeatureProcessor
This change adds early returns in EditorModeFeatureProcessor to stop a crash in the material editor. The crash is not happening on development for the automated testing project. It was happening on a game jam project, as described in this issue. https://github.com/o3de/o3de/issues/11382

These changes resolved the crash but there should be deeper investigation done in conjunction with owners of EditorModeFeatureProcessor.

Signed-off-by: gadams3 <guthadam@amazon.com>