Fixed the constraint checks on common_view
common_view was checking if the View was copyable, where it should have been checking if the iterator type of the view was copyable.

Fixed the range based container constructors and *_range functions to forward the container to the views::common adapter.

Added an implementation of the C++23 `to_underlying_type` for casting anenum vlaue to its underlying type.

Updated the `AZStd::ranges::views::join_with_view` and `AZStd::common_iterator`, to not use an AZStd::variant when compiling C++17, as it can't be used in a constexpr context.
Under the hood the `AZStd::variant` uses a placement new, which pre-C++20 can't be used in constexpr, to emplace the union alternative.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>