Motion Matching: Added standard scaler
The standard scaler can be used to normalize the feature matrix, the query vector and other data. It standardizes features by removing the mean and scaling to the unit variance. This implementation is mimicing the behavior of the standard scaler from scikit-learn (sklearn.preprocessing.StandardScaler). As we use floats by default, our errors are a bigger, especially if the variance is small as this leads to a division by a small value. In case the calculated standard deviation for a given feature is smaller than the given s_epsilon value, the standard deviation gets force-set to 1.0 to avoid divisions by infinity and preserve the value when doing the transform -> inverse-transform roundtrip.

Signed-off-by: Benjamin Jillich <jillich@amazon.com>