Fix the symbol visibility of the ITrace methods for Android (#11710)
The whole ITrace class needs to be exported for the symbols to resolve
correctly on Android. The `static ITrace*` that keeps the ITrace
instance must come only from the O3DEKernel library, and exporting the
whole class seems to be required to get this to work for Android. Methods
that must be in the O3DEKernel library are now defined in a separate
implementation file.

Signed-off-by: Chris Burel <burelc@amazon.com>