Support frustum shape in AuxGeom (#11728)
This PR adds support for AZ::Frustum in AuxGeom using code originally written for culling. The original code has been cleaned up an corrected a little but remains largely the same.

As part of this work, the ability to retrieve the corners of a frustum was moved into the AZ::Frustum class itself, along with unit tests to validate the new functionality.