Account for block shifts in total allocated bytes
When finding an insertion point in the tree for an allocation, the search
first finds a block that is big enough to hold the allocation. If that
block is too big, the block is split into two blocks, and only one is used.
However, that split only happens when there's enough room for the new
`block_header` and the new `free_node`. Further complicating this is
alignment. If the alignment is so high that there's room for a new
block/free_node, the block is split. But if the alignment is just right,
such that the offset necessary to achieve the alignment is less than the
size of the block_header and free_node (between 0 and 56 bytes), then
instead of splitting the block, the block is _shifted_. In order to shift
the start of the block higher, the block's _previous_ block (because
they're in a doubly-linked list) must be updated to point to the new start
of the shifted block. This necessarily increases the size of that previous
block.

Now, that block size is used to track the current total number of bytes
allocated. When an allocation occurs, `mTotalAllocatedSizeTree` is
incremented by the value of the block size. When a deallocation occurs, the
opposite happens. The trouble comes in when the block's size has mutated
between the time of allocation and deallocation. The tracking information
is decremented by the shifted size, which subtracts more than was
allocated.

This change updates the allocated size to correctly track the in-use bytes
that result from shifting a block.

Signed-off-by: Chris Burel <burelc@amazon.com>