Updated the Meshlets setup instructions to address problems I encountered while setting it up myself.
I also removed several unnecessary files from the ASV_GPU_and_CPU_Demo folder. These served as examples of what the files would look like after applying the steps in the Readme file, but there weren't actually referenced in the Readme, and they go stale quickly as changes continue to be made in AtomSampleViewer. It will be less confusing for users to just rely on the guidance in the Readme, and the folder only contains those files that the user should actually copy into AtomSampleViewer.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>