Adding implementation of C++20 civil calendar and hh_mm_ss
Updated AZStd::chrono implementation to alias the std::chrono
implementation for any C++17 or below names.

For the C++20 date facilities, the
[time.cal](http://eel.is/c++draft/time#cal) was implemented to allow
accessing calendar dates

The [time.hms](http://eel.is/c++draft/time#hms) is used to represent a
time of day.

An implementation of a utc_clock
[time.clock.utc](http://eel.is/c++draft/time#clock.utc) was added to
allow querying UTC time.

All this changes are for allowing portable query of a UTC timestamp with
fractional seconds that supports a precision on the order of microseconds.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>