Removing the `ApplicationTest.MakePathAssetRootRelative_AbsPath_RootLowerCase_Valid` unit test.
The test is invalid as it is assuming that the asset root path is lowercase.
This is incorrect as only the files within the asset root is lowercase, while the files outside of it retain their case.

The issue with this test becomes apparent on Linux where a temporary directory is used for the asset root.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>