Initial checkin of prototype builders for the material pipeline abstraction.
The new MaterialTypePreBuilder reads a materialtype[ex] file (the ex is temporary, it won't be part of the final design), "generates" necessary .shader and .azsl files, as well as an intermediate .materialtype file to reference the shaders. Then the normal material and shader builders take over. I had no problem referencing the final assets in material editor and O3DE level editor. Touching the original materialtype(ex) did not cause recompilation of the shader code, as expected.

The next step is to add source dependencies so that touching the "template" files (MinimalPBR shaders) that are used to "generate" the intermediate shaders will trigger the MaterialTypePreBuilder to run. This will be necessary for testing hot reload of the final materials.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>