Simplified the EventFactory management of the Event Loggers container
Previously it was using a convoluted logic around having a second vector
contain indexes into the event logger pointer vector in order to implement a binary
search.
This has problems as the implementation would have to perform a linear
iteration over the entire index vector in order to update the index
values.

This logic wasn't implemented correctly, resulting in the
FindEventLogger failing to locate an event logger that was actually
registered due to the logger id -> index vector containing stale
indices.

Address feedback provided in PR around documenting the EventFactory
inteface.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>