Make allocator instances always available, before `main()` starts
This changes the lifetime of the allocator instances to outlive any
variable that uses them. Allocators are now created on demand, and
destroyed at program exit.

Allocators that use other allocators internally should call
`AZ::AllocatorInstance<SubAllocator>::Get()` in their constructor, to
ensure the correct destruction order of the suballocator.

The OSAllocator now tracks allocations, to provide better visibility into
potential memory leaks.

The tests have been changed to check for leaks directly, instead of
checking for leaks in the allocator's destructor. This has been done to
accommodate the allocator's new lifetime.

Some of the existing API has been kept, like the `Destroy` method, even
though it is now a no-op. This was done to minimize the diff for the
initial change.

Signed-off-by: Chris Burel <burelc@amazon.com>