CMake referenced Gems external_subdirectories fixes
This is to fix a limitation with external_subdirectories recurse logic when
gathering the list of external_subdirectories to visit for any gems
referenced indirectly in the project.json/engine.json via the
"gem_names" field.

The limitation was that only the gem paths of the referenced "gem_names"
had cmake `add_subdirectory` command called on them.
If thoses gems contained additional external subdirectories referenced
in the "external_subdirectories" field in the gem.json, those
subdirectories were NOT added to CMake via `add_subdirectory`

The change here is to create a mapping of <gem-name> (via the
`O3DE_EXTERNAL_SUBDIRS_GEM_<gem_name>` property) to all external
subdirectories that are descendents for all gems.

Then when a gem is referenced by "gem_name", not only is the gem path
queried, but the external subdirectories associated with that gem as
well.

Next refactored the Subdirectories.cmake to remove the
duplicate logic for gathering the project, engine and o3de_manifest
external_subdirectories.

The functions that perform the gathering has been consolidated into
functions which are parameterized to except the o3de object type,
o3de object "<object>_name" value and the name of the json file for the
object type (i.e engine.json, project.json, o3de_manifest.json).

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>