Added the concept of a .materialpipeline file for arranging shader templates. Added material pipelines for MainPipeline and LowEndPipeline. There are listed by a new setreg file that MaterialTypePreBuilder loads.
Duplicated the MinimalPbr forward pass shader, and the common depth and shadowmap shaders to populate the templates for these material pipelines. The only change I made was in LowEndPipeline/ForwardPass.azsli, I #defined the low end preprocessor flags that normally appear in ShaderQualityOptions.azsli, and changed the final output code to pack the appropriate render targets directly instead of checking UNIFIED_FORWARD_OUTPUT.

At some point we'll probably want to factor out the depth and shadowmap shaders, as they are exact duplicates at this point, but that can wait for later.

The MaterialTypePreBuilder now utilizes the templates from the material pipeline, instead of copying files from hard-coded paths. It still doesn't actually stitch any code from the material type, it just uses the template azsl files as-is. It's just a proof of concept. A future revision will actually stitch some code together.

In order to support multiple pipelines, for now the MaterialTypePreBuilder outputs two .materialtype files, and you have to pick the one you want when creating a new material. This is the next thing I'll work on fixing, to make one MaterialTypeAsset support both pipelines.

I deviated from the original design a bit here in a couple ways. I realized it would be cleaner to have the .materialpipline file list both the template shader and template azsl, instead of relying on the shader to reference the azsl file and then replace that string. Also, instead of copying content from the template azsl file into the intermediate folder, it will be simpler if we can make an azsl file in the intermediate folder that just #includes the template azsl file. That way, the MaterialTypePreBuilder doesn't have to preprocess the template azsl and report them as dependencies. We can let the shader asset builder do all that heavy lifting when it finds the intermediate azsl that references the original.

Also fixed a bug in StringFunc::ReplaceExtension.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>